refactor: patch 함수 위치 수정
