+++ android/app/src/main/java/com/zzang/chongdae/data/remote/source/OfferingDetailDataSourceImpl.kt
@@ -25,9 +25,4 @@ class OfferingDetailDataSourceImpl
         override suspend fun saveParticipation(participationRequest: ParticipationRequest): Result<Unit, DataError.Network> =
             safeApiCall { participationApiService.postParticipations(participationRequest) }
 
-        override suspend fun patchOffering(
-            offeringId: Long,
-            offeringModifyRequest: OfferingModifyRequest,
-        ): Result<OfferingDetailResponse, DataError.Network> =
-            safeApiCall { offeringApiService.patchOffering(offeringId, offeringModifyRequest) }
     }
