+++ android/app/src/main/java/com/zzang/chongdae/data/source/offering/OfferingRemoteDataSource.kt
@@ -2,9 +2,11 @@ package com.zzang.chongdae.data.source.offering
 
 import com.zzang.chongdae.common.handler.DataError
 import com.zzang.chongdae.common.handler.Result
+import com.zzang.chongdae.data.remote.dto.request.OfferingModifyRequest
 import com.zzang.chongdae.data.remote.dto.request.OfferingWriteRequest
 import com.zzang.chongdae.data.remote.dto.response.offering.FiltersResponse
 import com.zzang.chongdae.data.remote.dto.response.offering.MeetingsResponse
+import com.zzang.chongdae.data.remote.dto.response.offering.OfferingDetailResponse
 import com.zzang.chongdae.data.remote.dto.response.offering.OfferingsResponse
 import com.zzang.chongdae.data.remote.dto.response.offering.ProductUrlResponse
 import com.zzang.chongdae.data.remote.dto.response.offering.RemoteOffering
@@ -29,4 +31,9 @@ interface OfferingRemoteDataSource {
     suspend fun fetchFilters(): Result<FiltersResponse, DataError.Network>
 
     suspend fun fetchMeetings(offeringId: Long): Result<MeetingsResponse, DataError.Network>
+
+    suspend fun patchOffering(
+        offeringId: Long,
+        offeringModifyRequest: OfferingModifyRequest,
+    ): Result<OfferingDetailResponse, DataError.Network>
 }
