+++ android/app/src/main/java/com/zzang/chongdae/domain/repository/OfferingRepository.kt
@@ -2,9 +2,11 @@ package com.zzang.chongdae.domain.repository
 
 import com.zzang.chongdae.common.handler.DataError
 import com.zzang.chongdae.common.handler.Result
+import com.zzang.chongdae.data.remote.dto.request.OfferingModifyRequest
 import com.zzang.chongdae.domain.model.Filter
 import com.zzang.chongdae.domain.model.Meetings
 import com.zzang.chongdae.domain.model.Offering
+import com.zzang.chongdae.domain.model.OfferingDetail
 import com.zzang.chongdae.domain.model.OfferingWrite
 import com.zzang.chongdae.domain.model.ProductUrl
 import okhttp3.MultipartBody
@@ -28,4 +30,9 @@ interface OfferingRepository {
     suspend fun fetchFilters(): Result<List<Filter>, DataError.Network>
 
     suspend fun fetchMeetings(offeringId: Long): Result<Meetings, DataError.Network>
+
+    suspend fun patchOffering(
+        offeringId: Long,
+        offeringModifyRequest: OfferingModifyRequest,
+    ): Result<OfferingDetail, DataError.Network>
 }
