+++ android/app/src/main/res/layout/fragment_offering_modify_optional.xml
@@ -266,7 +266,7 @@
                 android:layout_marginBottom="@dimen/size_34"
                 android:background="@drawable/btn_background_selector"
                 android:fontFamily="@font/suit_medium"
-                android:onClick="@{() -> vm.postOffering()}"
+                android:onClick="@{() -> vm.postOfferingModify()}"
                 android:text="@string/write_start_offering"
                 android:textColor="@color/white"
                 android:textSize="@dimen/size_16"
